// shared-library/src/lib/models/myrefunds/request/RemboursementRequest.ts
export class RemboursementRequest {
  numParticipant: string = "";
  dateDeRemboursement1: string = "";
  dateDeRemboursement2: string = "";
  nombreDeLignes: string = "";
  montant: number = 0;
  datePrevoyance: string = "";
  natureDeSoins: string = "";
  beneficiaire: string = "";
  decompteParticipant: string = "";
  dateEmission: string = "";
  decompteCode: string = "";
  envoiemail: string = "";
  emailSender: string = "";
  isSiteWeb: boolean = false;
  platform: string = "";
  browser: string = "";
  engine: string = "";
  numeroBordereau: string = "";
  codeBordereau: string = "";

  constructor(init?: Partial<RemboursementRequest>) {
    Object.assign(this, init);
  }
}